[FIX] stock_landed_costs: additional costs with multiple UoMs
When making a landed costs with pickings using
unit of measures different than the default products unit of measure,
the additional costs added to the quants inventory value were wrong.

This is due to the fact the quants unit of measure is always
the quants product unit of measure, while the quantity
in the adjustment lines were assumed in the stock moves
unit of measure.

As there is no unit of measure field on the adjusment lines,
it makes sense to assume the quantity within the product
unit of measure, to display the quantity within the unit
of measure the user uses the most, which is, normally,
the product unit of measure.

Thanks to this, the additional costs added to the quants
inventory value is now correct.

opw-668374