[FIX] product: pricelist item delete cascade on pricelist deletion
On deletion of a pricelist,
the pricelist items were not deleted.

Meaning, if you deleted a pricelist,
and this pricelist was used in the product pricing,
the line with the pricing remained, but without any
pricelist set. Besides, as the field is required
in the view, this prevented to edit anything without
correcting the pricing line first.

opw-669652