[FIX] point_of_sale: bank statement starting balance
When creating a new pos.session, the starting balance of created bank
statements where all the sames.

By creating bank statements with the same `journal_id` in context, they
all have the starting balance for this journal.

Create them manually to force having the starting balance for the
statement's journal.