[FIX] account: better error message when emptying journal account number
The account number field on a journal is a related field to the
number of the bank account associated to the journal.

As the number of a bank account is a required field,
any attempt to empty it from the journal fails.

Besides, unlinking the bank account from the journal when emptying
the account number on the journal form doesn't seem to be
the appropriate behavior. Imagine you give an account number
to the journal, then empty it, then re-fill it, you probably
do not want the journal entries to be associated to a new bank account,
you probably wanted the same bank account, but renamed, while
it would be the case if we simply unlink it.

If you would like to delete the bank account, you have
to do it from the bank accounts list itself.

opw-670472