[FIX] sale_timesheet: remove unused, invisible `timesheet_ids`
The `timesheet_ids` was added in the `sale.order` form.

As it was invisible, and not used by the attrs of another
field, or by an onchange, this is pointless to load
it for no reasons.

Besides, this could lead to issues regarding the form validation,
in case the analytic lines were not valid.
e.g. if an editable list view is used to load
these analytic lines and if a field in this editable list
has been made required,
while analytic lines exist with this field not filled.
In such a case, it was not possible for the user to correct the
analytic lines, as they were invisible, and it was not the place
to correct them anyway.

opw-672184