[FIX] mrp_repair: `tax_id` is a many2many
This is related to 27890a93aadce606a7c6c826d7c78e16162832cd

`compute_all` indeed does expect ids, and not browse records,
but it expects a list of ids for the `ids` parameter, or a single
id, but not `False`.

Besides, `line.tax_id` is a many2many, despite the fact
its naming seems to say otherwise, and therefore `.ids`
should be used in order to pass all the taxes of the line,
not just one.