[FIX] point_of_sale: correct sign of fixed tax
Oversight of 55d97d3bc120668e6bbffa2d7d702354242a7253. Although the
previous commit gave the correct behaviour for a negative fixed tax with
a base_amount of 0, it was still incorrect for positive fixed taxes with
a base_amount of 0.

This rewrites the code to more closely resemble the relevant code of
account.tax._compute_amount() solving the above mentioned issue.