[FIX] crm: default subscription
When a CRM activity is defined, the subscription is activated by default
for any user. There is no way to change it since the field is not in
the form view. Therefore, a customer is automatically subscribed to any
internal activity, and the user must manually remove the subscription
afterwards.

The `default` field is displayed in the forw view, so this default
behavior can now be configured for a given activity.

opw-672453